Remove redundant serialport drivers
The original version of node-dmx includes drivers for various serial port DMX devices. All of these drivers use the `serialport` library, which itself uses binaries which often need compiling when they are installed. This adds a significant overhead when used for example in a Docker container, since all of the tooling to compile the binaries must also be added into the container.

However, since we only use Artnet devices in the museum, the serial port devices are redundant, and can be safely removed from this fork.

This commit removes the redundant drivers and `serialport` dependancies.